+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/interop/values/BoxedTestValue.java
@@ -45,6 +45,10 @@ public final class BoxedTestValue implements TruffleObject {
         this.value = value;
     }
 
+    public Object getValue() {
+        return value;
+    }
+
     @Override
     public ForeignAccess getForeignAccess() {
         return ForeignAccess.create(BoxedTestValue.class, new StandardFactory() {